+++ test/modules/datatype/int.test.ts
@@ -1,61 +1,72 @@
-import { modules } from "../../../src";
+import { ChacaError, modules } from "../../../src";
 import { describe, expect, it } from "vitest";
 
-describe("# Datatype Int Number test", () => {
-  const ARRAY_LENGTH_TEXT = 1000;
+describe("datatype.int", () => {
+  const VALUES_LIMIT = 1000;
 
-  it("Passing a max value (5). Should return a int number less than 5", () => {
-    const values = Array.from({ length: ARRAY_LENGTH_TEXT }).map(() =>
-      modules.datatype.int({ max: 5 }),
-    );
+  describe("max and min argument", () => {
+    it("max = 20 & min = 20. should return 20", () => {
+      for (let i = 0; i < VALUES_LIMIT; i++) {
+        const value = modules.datatype.int({ max: 20, min: 20 });
 
-    expect(values.every((v) => v <= 5 && Number.isInteger(v))).toBe(true);
+        expect(value).toBe(20);
+      }
+    });
+
+    it("min = 0 & max = 0. should return 0", () => {
+      for (let i = 0; i < VALUES_LIMIT; i++) {
+        const value = modules.datatype.int({ max: 0, min: 0 });
+
+        expect(value).toBe(0);
+      }
+    });
+
+    it("max = -1 & min = -10. should return an integer number between -1 and -10", () => {
+      for (let i = 0; i < VALUES_LIMIT; i++) {
+        const value = modules.datatype.int({ max: -1, min: -10 });
+
+        expect(value).toBeGreaterThanOrEqual(-10);
+        expect(value).toBeLessThanOrEqual(-1);
+        expect(Number.isInteger(value)).toBe(true);
+      }
+    });
+
+    it("min = 5 & max = 0. should throw an error", () => {
+      expect(() => modules.datatype.int({ min: 5, max: 0 })).toThrow(
+        ChacaError,
+      );
+    });
+
+    it("max = 100 & min = 0. should return an integer number between 0 and 100", () => {
+      for (let i = 0; i < VALUES_LIMIT; i++) {
+        const value = modules.datatype.int({ max: 100, min: 0 });
+
+        expect(value).toBeGreaterThanOrEqual(0);
+        expect(value).toBeLessThanOrEqual(100);
+        expect(Number.isInteger(value)).toBe(true);
+      }
+    });
   });
 
-  it("passing 20 as max and min argument. Sould return 20", () => {
-    const values = Array.from({ length: ARRAY_LENGTH_TEXT }).map(() =>
-      modules.datatype.int({ max: 20, min: 20 }),
-    );
+  describe("max argument", () => {
+    it("max = 5. should return a int number less than 5", () => {
+      for (let i = 0; i < VALUES_LIMIT; i++) {
+        const value = modules.datatype.int({ max: 5 });
 
-    expect(values.every((v) => v === 20 && Number.isInteger(v))).toBe(true);
+        expect(value).toBeLessThanOrEqual(5);
+        expect(Number.isInteger(value)).toBe(true);
+      }
+    });
   });
 
-  it("Passing min: 0 && max: 100. Should return an int number between 0 and 100", () => {
-    const values = Array.from({ length: ARRAY_LENGTH_TEXT }).map(() =>
-      modules.datatype.int({ max: 100, min: 0 }),
-    );
+  describe("min argument", () => {
+    it("min = -50. should return an int grater than -50", () => {
+      for (let i = 0; i < VALUES_LIMIT; i++) {
+        const value = modules.datatype.int({ min: -50 });
 
-    expect(values.every((v) => v <= 100 && v >= 0 && Number.isInteger(v))).toBe(
-      true,
-    );
-  });
-
-  it("Passing min: -10 && max: -1. Should return an int number between -1 and -10", () => {
-    const values = Array.from({ length: ARRAY_LENGTH_TEXT }).map(() =>
-      modules.datatype.int({ max: -1, min: -10 }),
-    );
-
-    expect(
-      values.every((v) => Number.isInteger(v) && v >= -10 && v <= -1),
-    ).toBe(true);
-  });
-
-  it("Passing min: -50. Should return an int grater than -50", () => {
-    const values = Array.from({ length: ARRAY_LENGTH_TEXT }).map(() =>
-      modules.datatype.int({ min: -50 }),
-    );
-
-    expect(values.every((v) => Number.isInteger(v) && v >= -50)).toBe(true);
-  });
-
-  it("Pass min=0 & max=0. Should return 0", () => {
-    const values = Array.from({ length: ARRAY_LENGTH_TEXT }).map(() =>
-      modules.datatype.int({ min: 0, max: 0 }),
-    );
-
-    for (const v of values) {
-      expect(Number.isInteger(v)).toBe(true);
-      expect(v).toBe(0);
-    }
+        expect(value).toBeGreaterThanOrEqual(-50);
+        expect(Number.isInteger(value)).toBe(true);
+      }
+    });
   });
 });