+++ test/modules/datatype/number.test.ts
@@ -1,48 +1,60 @@
-import { modules } from "../../../src";
+import { ChacaError, modules } from "../../../src";
 import { describe, expect, it } from "vitest";
 
-const TEST_COUNT_VALUES = 1000;
+const VALUES_LIMIT = 1000;
 
-describe("# Datatype Number test", () => {
-  it("Without arguments. Should return float numbers", () => {
-    const allValues = Array.from({ length: TEST_COUNT_VALUES }).map((v) =>
-      modules.datatype.number(),
-    );
+describe("datatype.number", () => {
+  it("no arguments. should return a integer or float number", () => {
+    for (let index = 0; index < VALUES_LIMIT; index++) {
+      const value = modules.datatype.number();
 
-    expect(allValues.every((v) => typeof v === "number")).toBe(true);
+      expect(typeof value === "number").toBe(true);
+    }
   });
 
-  it("With min argument", () => {
-    const allValues = Array.from({ length: TEST_COUNT_VALUES }).map(() =>
-      modules.datatype.number({ min: 5 }),
-    );
+  describe("min argument", () => {
+    it("min = 5. should return a number greater than 5", () => {
+      for (let index = 0; index < VALUES_LIMIT; index++) {
+        const value = modules.datatype.number({ min: 5 });
 
-    expect(allValues.every((v) => v >= 5)).toBe(true);
+        expect(value).toBeGreaterThanOrEqual(5);
+      }
+    });
   });
 
-  it("With max argument", () => {
-    const allValues = Array.from({ length: TEST_COUNT_VALUES }).map((v) =>
-      modules.datatype.number({ max: 20 }),
-    );
+  describe("max argument", () => {
+    it("max = 5. should return a number less than 5", () => {
+      for (let index = 0; index < VALUES_LIMIT; index++) {
+        const value = modules.datatype.number({ max: 5 });
 
-    expect(allValues.every((v) => v <= 20)).toBe(true);
+        expect(value).toBeLessThanOrEqual(5);
+      }
+    });
   });
 
-  it("With min and max argument", () => {
-    const allValues = Array.from({ length: TEST_COUNT_VALUES }).map(() =>
-      modules.datatype.number({ max: 1000, min: -1000 }),
-    );
+  describe("max and min argument", () => {
+    it("max = 1000 & min = -1000. should return a number between -1000 and 1000", () => {
+      for (let index = 0; index < VALUES_LIMIT; index++) {
+        const value = modules.datatype.number({ max: 1000, min: -1000 });
 
-    expect(allValues.every((v) => v <= 1000 && v >= -1000)).toBe(true);
-  });
+        expect(value).toBeLessThanOrEqual(1000);
+        expect(value).toBeGreaterThanOrEqual(-1000);
+      }
+    });
 
-  it("With min=-1000, max=1000, precision=0. Should return a integer number between -1000 and 1000", () => {
-    const allValues = Array.from({ length: TEST_COUNT_VALUES }).map(() =>
-      modules.datatype.number({ max: 1000, min: -1000, precision: 0 }),
-    );
+    it("max = 0 & min = 0. should return a 0", () => {
+      for (let index = 0; index < VALUES_LIMIT; index++) {
+        const value = modules.datatype.number({ max: 0, min: 0 });
 
-    expect(
-      allValues.every((v) => v <= 1000 && v >= -1000 && Number.isInteger(v)),
-    ).toBe(true);
+        expect(Number.isInteger(value)).toBe(true);
+        expect(value).toBe(0);
+      }
+    });
+
+    it("min = 10 & max = 5. should throw an error", () => {
+      expect(() => modules.datatype.number({ min: 10, max: 5 })).toThrow(
+        ChacaError,
+      );
+    });
   });
 });